# ⚡️ Speed up function `_estimate_string_tokens` by 221%